(PC-5221) move tabnav and rename its components
