+++ src/features/navigation/RootNavigator.tsx
@@ -2,8 +2,8 @@ import { NavigationContainer, NavigationContainerRef } from '@react-navigation/n
 import { createStackNavigator } from '@react-navigation/stack'
 import React from 'react'
 
-import { RootTabNavigator } from './RootTabNavigator'
 import { onNavigationStateChange } from './services'
+import { TabNavigator } from './TabBar/TabNavigator'
 
 export type RootStackParamList = {
   TabNavigator: undefined
@@ -16,7 +16,7 @@ export const RootNavigator: React.FC = () => {
   return (
     <NavigationContainer onStateChange={onNavigationStateChange} ref={navigationRef}>
       <RootStack.Navigator initialRouteName="TabNavigator" screenOptions={{ headerShown: false }}>
-        <RootStack.Screen name="TabNavigator" component={RootTabNavigator} />
+        <RootStack.Screen name="TabNavigator" component={TabNavigator} />
       </RootStack.Navigator>
     </NavigationContainer>
   )
